MAINT: special.wrightomega: migrate to xsf
