+++ scipy/special/xsf_wrappers.h
@@ -150,6 +150,9 @@ double cephes_polevl_wrap(double x, const double coef[], int N);
 double special_wright_bessel(double a, double b, double x);
 double special_log_wright_bessel(double a, double b, double x);
 
+double xsf_wrightomega(double z);
+npy_cdouble xsf_cwrightomega(npy_cdouble z);
+
 double xsf_beta(double a, double b);
 double xsf_betaln(double a, double b);
